Simplified fast lisp and aligned Rust
This simplifies the fast lisp implementation by removing the need for
iterating through to count the number of factor pairs. This was done so
we could create a stack allocated byte array, which we need to know the
size of first. In practice though this means we are computing the factor
pairs twice.

Instead we can make this assumption:
1. In practice other than the zero product case, I've only ever seen two
factor pairs for a palindrome product. We can make a buffer allocated
for three factor pairs to be on the safe side, count factor pairs as we
go, and then copy into a correctly sized byte array.

I made the Rust implementation mirror this a bit closer, using VecArray
to ensure that when it collect factor pairs, it does it in a way that's
also stack allocated.

This updates the benchmarks accordingly. There's a small performance
gain in both implementations, but quite small.