Use better runner logic for haskell and mirror
Haskell was hoisting iterations for any unique set of inputs, so if we
were going to do 5k runs with a certain set of inputs, it would do it
only once, and then use determinism to fold into the accumulator, which
is brilliant, but made the benchmarks unreliable.

This changes the runner logic and finally avoid GHC hoisting - now
Haskell performs worse of any language, although is still relatively
fast. This is expected performance at this point.