Don't handle zero case and update README
Handling the zero product case involves extra branching for a case we
don't even benchmark. This updates all solutions to remove handling the
zero edge case, and updates the README to indicate that we only support
strictly positive inputs.

I also changed benchmarks to range from [2, 999] to get rid of cases
where we have to generate factors with 1 * a large number of inputs, as
this seems like an odd edge case as well that we might as well remove
entirely from benchmarking