Several improvements to Common Lisp
This adds more efficient server argument parsing - unlikely to yield
benchmark differences, but a nice improvement. It also adds a
with-divmod macro to help clean up call sites where we
multiple-value-bind on truncate (quot + rem in one op).

It also adds a helper to sum the factors vector for the accumulator
step. The reduce method we used was clean, but requires generic dispatch
which is generally slow. This is only once for each iteration after
finding the correct palindrome and factors, so it's unlikely to yield
much gain, but nice to tidy this up regardless.