Align Coalton to Common Lisp runners
This updates the Coalton runners to use the same new macro for more
efficient line parsing in the server, and updates the disassemble script
to show the top runner loops themselves. This verifies that there's no
generic math, and the only boxing occurs when a tuple is returned - this
is unfortunately unavoidable with idiomatic Coalton.