Simplify all cmd parsers for benchmark consistency
This drastically simplifies all stdin cmd parsers. This now expects
strictly correct input, and essentially does no validation or error
handling. We expect that because we control benchmarking entirely, that
we can deliver the correct input, and simplify parsing to keep overhead
as low as possible, and deliver more consistent and accurate benchmark
results.