Manual while loops are better optimized by LLVM
Who would have thought? This closes most of the performance gap between
iterative and functional. The last piece is that factor pair collection
is auto vectorized in the functional version - next step is to figure
out if there's a way to create a more vectorizable iterative version of
factor pair collection.