These can just be ifs
We don't need while loops for these. They can just be an if chain - we
know our scratch capacity is at most 16, but in reality we will never
hit 16, because if we are ever >= SIMD_WIDTH, we process a full lane
unconditionally, this means we will always be < 16 strictly, so we can
have s imply cascade without loop overhead here.