Switch product stepping to add/sub and benchmarks
This switches product stepping in rust to use add/sub in row
searching/looping, since we know for each step we are only adding or
substracting x effectively, we can remove a mul op in the hot path.

This also updates the functional rust solution to use our new magic
multiplier div_rem that works for numbers up to 999 (which is great for
our problem domain).

Updates benchmarks as well.