This is more efficient form of pal checking
We already check if n is a multiple of 10 to get rid of trailing zeros.
They aren't palindromes, and they mess up our ability to do our half
reversal check.

Anytime we write div and mod of the same fixed divisor, LLVM full fuse
it into a single efficient instruction set using a magic constant.

That means that we can do the first peel, and use the remainder to early
exit on trailing zeros, which saves us a peel later.