Updates simd for faster largest
This updates simd to use a different algorithm for largest, which
drastically speeds up performance on largest searches. I tried the same
strategy for smallest, but it drastically slowed down smallest. These
two searches are fundamnetally different, so I'm maintaining different
code paths for each, so each search performs as well as it can.